Moved DemoCourse to search.tests.utils
This is a shared class. It should not live in a single test module.
