+++ search/tests/test_course_discovery_views.py
@@ -1,10 +1,10 @@
 """ High-level view tests"""
 from django.test.utils import override_settings
 
+from search.tests.test_views import MockSearchUrlTest
 from search.tests.tests import TEST_INDEX_NAME
-from .test_views import MockSearchUrlTest
-from .test_course_discovery import DemoCourse
-from search.tests.utils import post_discovery_request
+from search.tests.utils import post_discovery_request, DemoCourse
+
 
 # Any class that inherits from TestCase will cause too-many-public-methods pylint error
 # pylint: disable=too-many-public-methods
